PDF-948 major code clean up for JDK 1.8 as well as javadoc clean up.

+++ icepdf/core/core-awt/src/main/java/org/icepdf/core/pobjects/graphics/RasterOps/ImageMath.java
@@ -270,7 +270,7 @@ public class ImageMath {
      *
      * @param x   the X interpolation parameter 0..1
      * @param y   the y interpolation parameter 0..1
-     * @param rgb array of four ARGB values in the order NW, NE, SW, SE
+     * @param p array of four ARGB values in the order NW, NE, SW, SE
      * @return the interpolated value
      */
     public static int bilinearInterpolate(float x, float y, int[] p) {
@@ -317,7 +317,7 @@ public class ImageMath {
     /**
      * Return the NTSC gray level of an RGB value.
      *
-     * @param rgb1 the input pixel
+     * @param rgb the input pixel
      * @return the gray level (0-255)
      */
     public static int brightnessNTSC(int rgb) {
@@ -404,7 +404,7 @@ public class ImageMath {
                 break;
         if (span > numKnots - 3)
             span = numKnots - 3;
-        float t = (float) (x - xknots[span]) / (xknots[span + 1] - xknots[span]);
+        float t = (x - xknots[span]) / (xknots[span + 1] - xknots[span]);
         span--;
         if (span < 0) {
             span = 0;
@@ -554,7 +554,7 @@ public class ImageMath {
         for (j = 0; j < length; j++) {
             while (out[i + 1] < j)
                 i++;
-            in[j] = i + (float) (j - out[i]) / (out[i + 1] - out[i]);
+            in[j] = i + (j - out[i]) / (out[i + 1] - out[i]);
         }
         in[length] = length;
